Remove print statements that are python-2 specific.
I don't think they are needed at all, if there's an exception on import, that should be sufficient to produce an OS error code.  And that way, the code is simpler than trying to make the code python 2/3 compatible with escaped parens and all that :)

I tested it on my local install with the changes and it's fine, it detected at first a missing pycurl, and then worked once I had it.